requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-kzt-exchangerates==0.0.2
requests==2.25.1
selenium==3.141.0
urllib3==1.26.2
